Fix: Split network concerns from express layer for better testing
