+++ api.js
@@ -1,0 +1,15 @@
+require('dotenv').config();
+const express = require('express');
+const bodyParser = require('body-parser');
+const routes = require('./routes');
+const { port } = require('./config');
+const cors = require('cors');
+
+const app = express();
+
+app.use(cors());
+app.use(bodyParser.urlencoded({ extended: true }));
+app.use(bodyParser.json());
+routes(app);
+
+module.exports = app;
