+++ index.js
@@ -1,16 +1,5 @@
-require('dotenv').config();
-const express = require('express');
-const bodyParser = require('body-parser');
-const routes = require('./routes');
+const app = require('./api');
 const { port } = require('./config');
-const cors = require('cors');
-
-const app = express();
-
-app.use(cors());
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(bodyParser.json());
-routes(app);
 
 app.listen(port, () => {
   console.log(`running on ${port}`);
